refactor(ver2): split grade scoring helpers for code health
