+++ src/utils/verificationV2/gradeContract.ts
@@ -116,80 +116,13 @@ export interface PackageGrade {
   generatedAt: string;
 }
 
-/** Clamps a finite value into the inclusive [min, max] range; non-finite input becomes min. */
-export const clamp = (value: number, min = 0, max = 1): number => {
-  if (!Number.isFinite(value)) {
-    return min;
-  }
-  return Math.min(max, Math.max(min, value));
-};
-
-/** Rounds a 0–100 grade to a single decimal place. */
-export const roundGrade = (grade: number): number => Math.round(grade * 10) / 10;
-
-/** Maps a 0–100 grade to a letter. A≥90 B≥80 C≥70 D≥60 F<60. */
-export const scoreToLetter = (grade: number | null): LetterGrade | null => {
-  if (grade === null || Number.isNaN(grade)) {
-    return null;
-  }
-  if (grade >= 90) return 'A';
-  if (grade >= 80) return 'B';
-  if (grade >= 70) return 'C';
-  if (grade >= 60) return 'D';
-  return 'F';
-};
-
-/**
- * Combines component scores into a 0–100 product grade, renormalizing the
- * weights over only the applicable (non-N/A) components. Returns null when no
- * component is applicable.
- */
-export const composeComponents = (components: ComponentScore[]): number | null => {
-  let weightSum = 0;
-  let weighted = 0;
-
-  for (const component of components) {
-    if (!component.applicable || component.score === null || !Number.isFinite(component.score)) {
-      continue;
-    }
-    weightSum += component.weight;
-    weighted += component.weight * component.score;
-  }
-
-  if (weightSum === 0) {
-    return null;
-  }
-
-  return roundGrade((weighted / weightSum) * 100);
-};
-
-/** Builds a Not-evaluated component placeholder with a factual reason. */
-export const notEvaluatedComponent = (key: ComponentKey, detail: string): ComponentScore => ({
-  key,
-  label: COMPONENT_LABELS[key],
-  weight: COMPONENT_WEIGHTS[key],
-  score: null,
-  applicable: false,
-  detail,
-});
-
-/** Builds a scored component with normalized score and metrics. */
-export const scoredComponent = (
-  key: ComponentKey,
-  score: number,
-  detail: string,
-  metrics?: Record<string, number>
-): ComponentScore => {
-  const normalized = clamp(score);
-  return {
-    key,
-    label: COMPONENT_LABELS[key],
-    weight: COMPONENT_WEIGHTS[key],
-    score: normalized,
-    applicable: true,
-    detail,
-    metrics,
-  };
-};
+export {
+  clamp,
+  composeComponents,
+  notEvaluatedComponent,
+  roundGrade,
+  scoreToLetter,
+  scoredComponent,
+} from './gradeScoring';
 
 export { FORECAST_GRADE_FORMULA_VERSION };
